Update MetaTrader order processing and DTOs
Changed the type of 'Magic' property to long in DTO and updated 'OrderStatus' enum to include 'Executed'. Enhanced OrderProcessor with account ID configuration and modified order status handling for better accuracy. Replaced Console.WriteLine with Debug.WriteLine in HttpService for better debugging.

+++ MetaTraderWorkerService/Dtos/MetaTraderOpenTradeOrderRequestDto.cs
@@ -50,7 +50,7 @@ public class MetaTraderOpenTradeOrderRequestDto
     [JsonProperty("stopPriceBase")]
     public string? StopPriceBase { get; set; } // Base price for calculating SL/TP, e.g., "OPEN_PRICE"
 
-    [JsonProperty("magic")] public int? Magic { get; set; } // Magic number for MetaTrader
+    [JsonProperty("magic")] public long? Magic { get; set; } // Magic number for MetaTrader
 
     // [JsonProperty("expirationType")]
     // public string? ExpirationType { get; set; } // Expiration type, e.g., "ORDER_TIME_SPECIFIED"
